Move links to constants
