+++ desktop/packages/mullvad-vpn/src/renderer/components/Support.tsx
@@ -1,7 +1,7 @@
 import { useCallback } from 'react';
 import styled from 'styled-components';
 
-import { links } from '../../config.json';
+import { links } from '../../shared/constants';
 import { messages } from '../../shared/gettext';
 import { useAppContext } from '../context';
 import { useHistory } from '../lib/history';
