+++ desktop/packages/mullvad-vpn/src/shared/constants/links.ts
@@ -1,0 +1,8 @@
+// This should only contain links to the mullvad website
+// No links to other websites should be added
+export const links = {
+  purchase: 'https://mullvad.net/account/',
+  faq: 'https://mullvad.net/help/tag/mullvad-app/',
+  privacyGuide: 'https://mullvad.net/help/first-steps-towards-online-privacy/',
+  download: 'https://mullvad.net/download/vpn/',
+} as const;
